clean up leaderboard-trial plus code separation
