+++ src/hooks/useScores.js
@@ -1,0 +1,23 @@
+import { useState, useContext } from "react";
+
+// Call API and get the scores for a specfic UserID
+
+export const useScores = () => {
+  const [scores, setScores] = useState();
+
+  const getScores = async (userID) => {
+    fetch("https://siasky.dev/leaderboard/users?userPK=" + userID)
+      .then((response) => response.json())
+      .then((data) => {
+        if (data) {
+          setScores(data[0]);
+        }
+      });
+  };
+
+  const resetScores = () => {
+    setScores();
+  };
+
+  return [scores, getScores, resetScores];
+};
